[optimisation]: minor update
